requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
edge-tts==7.2.3
cloudinary==1.44.1
python-dotenv==1.1.1
pydantic==2.11.7
